+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 dd mm yyyy
 
-- Added API for alternative password login.
+- Added password login as an alternative to EVE SSO.
 - Added a css file with a minimal dark theme that plugins can use.
 - Removed support for build-in PHP development server.